+++ src/lib/brain/executive/index.ts
@@ -1,0 +1,21 @@
+export { isTravelExecutiveEnabled, TRAVEL_EXECUTIVE_FEATURE_ID } from './feature'
+export { buildExecutiveContext } from './contextBuilder'
+export { collectBudgetWarnings, executiveBudgetLine } from './budgetIntelligence'
+export {
+  detectRejectedDestinations,
+  learnRejectedDestinations,
+  applyRejectedDestinationsFilter,
+} from './rejectedDestinations'
+export {
+  detectOptimizationAxis,
+  optimizeDiscoveryRanking,
+} from './discoveryOptimizer'
+export { composeExecutiveDiscoveryReply } from './executiveResponseComposer'
+export { processExecutiveIntelligence, executiveReasoningSnapshot } from './executiveEngine'
+export type {
+  OptimizationAxis,
+  ExecutiveTravelStyle,
+  ExecutiveContext,
+  ExecutiveEnhancement,
+  ExecutiveProcessInput,
+} from './types'